requirements.txt
datasets
evaluate
numpy
peft
pytorch_crf
torch==2.0.1
transformers
seqeval